Refactor: move message classes to separate package
