+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/AttributeConstraint.java
@@ -1,11 +1,6 @@
 package org.molgenis.data.validation.constraint;
 
-public enum AttributeConstraint implements Constraint
+public enum AttributeConstraint
 {
-	COMPOUND_PARENT, DEFAULT_VALUE_EMAIL, DEFAULT_VALUE_ENTITY_REFERENCE, DEFAULT_VALUE_ENUM, DEFAULT_VALUE_HYPERLINK, DEFAULT_VALUE_MAX_LENGTH, DEFAULT_VALUE_TYPE, MAPPED_BY_REFERENCE, MAPPED_BY_TYPE, NAME, NON_COMPOUND_CHILDREN, ORDER_BY_REFERENCE, TYPE_UPDATE_BIDIRECTIONAL, TYPE_UPDATE;
-
-	public ConstraintType getType()
-	{
-		return ConstraintType.ATTRIBUTE;
-	}
+	COMPOUND_PARENT, DEFAULT_VALUE_EMAIL, DEFAULT_VALUE_ENTITY_REFERENCE, DEFAULT_VALUE_ENUM, DEFAULT_VALUE_HYPERLINK, DEFAULT_VALUE_MAX_LENGTH, DEFAULT_VALUE_TYPE, MAPPED_BY_REFERENCE, MAPPED_BY_TYPE, NAME, NON_COMPOUND_CHILDREN, ORDER_BY_REFERENCE, TYPE_UPDATE_BIDIRECTIONAL, TYPE_UPDATE
 }
