+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/AttributeValueConstraint.java
@@ -1,11 +1,6 @@
 package org.molgenis.data.validation.constraint;
 
-public enum AttributeValueConstraint implements Constraint
+public enum AttributeValueConstraint
 {
-	EMAIL, ENTITY_REFERENCE, ENUM, EXPRESSION, HYPERLINK, MAX_LENGTH, NOT_NULL, RANGE, READ_ONLY, TYPE, UNIQUE;
-
-	public ConstraintType getType()
-	{
-		return ConstraintType.ATTRIBUTE_VALUE;
-	}
+	EMAIL, ENTITY_REFERENCE, ENUM, EXPRESSION, HYPERLINK, MAX_LENGTH, NOT_NULL, RANGE, READ_ONLY, TYPE, UNIQUE
 }
