--- molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/Constraint.java
@@ -1,12 +1,0 @@
-package org.molgenis.data.validation.constraint;
-
-/**
- * Limitation placed on data.
- */
-public interface Constraint
-{
-	/**
-	 * Type of limitation placed on data (e.g. entity type, attribute, attribute value)
-	 */
-	ConstraintType getType();
-}
