--- molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/ConstraintType.java
@@ -1,14 +1,0 @@
-package org.molgenis.data.validation.constraint;
-
-/**
- * @see PackageConstraint
- * @see TagConstraint
- * @see EntityTypeConstraint
- * @see AttributeConstraint
- * @see AttributeValueConstraint
- * @see DefaultValueReferenceConstraint
- */
-public enum ConstraintType
-{
-	PACKAGE, ENTITY_TYPE, ATTRIBUTE, TAG, ATTRIBUTE_VALUE, ENTITY
-}
