+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/ConstraintViolation.java
@@ -2,7 +2,5 @@ package org.molgenis.data.validation.constraint;
 
 public interface ConstraintViolation
 {
-	Constraint getConstraint();
-
 	void accept(ConstraintViolationVisitor constraintViolationVisitor);
 }
