+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/EntityTypeConstraint.java
@@ -1,12 +1,6 @@
 package org.molgenis.data.validation.constraint;
 
-public enum EntityTypeConstraint implements Constraint
+public enum EntityTypeConstraint
 {
-	LOOKUP_ATTRIBUTES_EXIST, LABEL_ATTRIBUTE_EXISTS, ID_ATTRIBUTE_EXISTS, ID_ATTRIBUTE_TYPE, ID_ATTRIBUTE_UNIQUE, ID_ATTRIBUTE_NOT_NULL, ID_ATTRIBUTE_REQUIRED, BACKEND_EXISTS, HAS_ATTRIBUTES, ATTRIBUTES_UNIQUE, ATTRIBUTE_IN_PARENT, EXTENDS_NOT_ABSTRACT, NAME, LABEL_NOT_EMPTY, LABEL_NOT_WHITESPACE_ONLY, PACKAGE_NOT_SYSTEM;
-
-	@Override
-	public ConstraintType getType()
-	{
-		return ConstraintType.ENTITY_TYPE;
-	}
+	LOOKUP_ATTRIBUTES_EXIST, LABEL_ATTRIBUTE_EXISTS, ID_ATTRIBUTE_EXISTS, ID_ATTRIBUTE_TYPE, ID_ATTRIBUTE_UNIQUE, ID_ATTRIBUTE_NOT_NULL, ID_ATTRIBUTE_REQUIRED, BACKEND_EXISTS, HAS_ATTRIBUTES, ATTRIBUTES_UNIQUE, ATTRIBUTE_IN_PARENT, EXTENDS_NOT_ABSTRACT, NAME, LABEL_NOT_EMPTY, LABEL_NOT_WHITESPACE_ONLY, PACKAGE_NOT_SYSTEM
 }
