+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/EntityTypeConstraintViolation.java
@@ -15,7 +15,6 @@ public final class EntityTypeConstraintViolation implements ConstraintViolation
 		this.entityType = requireNonNull(entityType);
 	}
 
-	@Override
 	public EntityTypeConstraint getConstraint()
 	{
 		return entityTypeConstraint;
