+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/constraint/TagConstraintViolation.java
@@ -15,7 +15,6 @@ public class TagConstraintViolation implements ConstraintViolation
 		this.tag = requireNonNull(tag);
 	}
 
-	@Override
 	public TagConstraint getConstraint()
 	{
 		return tagConstraint;
