+++ molgenis-data-validation/src/main/resources/l10n/data_validation_en.properties
@@ -12,3 +12,7 @@ V10=Cannot delete entity ''{0}'' because entity ''{1}'' depends on it.
 V11=Value '{2}' for attribute '{1}' is referenced by entity '{0}'.
 V12="Entity type ''{0}'' attribute ''{1}'' cannot be unique and have a default value at the same time."
 V13="Entity type ''{0}'' attribute ''{1}'' cannot have an expression and have a default value at the same time."
+# Tag validation messages
+V20=Unknown relation IRI ''{1}'' for tag ''{0}''
+# Package validation messages
+V30=Modifying system package ''{0}'' is not allowed
